Week4_FastAPI/hw4_1/requirements.txt
fastapi[all]==0.115.5
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-dotenv==1.0.1